qmicli,dms: fix warnings with -Wsign-compare
  qmicli-dms.c: In function ‘get_stored_image_list_stored_images_ready’:
  qmicli-dms.c:2833:25: warning: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘guint’ {aka ‘unsigned int’} [-Wsign-compare]
   2833 |         if (image_index >= image->sublist->len) {
        |                         ^~

+++ src/qmicli/qmicli-dms.c
@@ -2830,7 +2830,7 @@ get_stored_image_list_stored_images_ready (QmiClientDms *client,
         if (image_index < 0)
             continue;
 
-        if (image_index >= image->sublist->len) {
+        if ((guint)image_index >= image->sublist->len) {
             g_printerr ("error: couldn't find '%s' image at index '%d'\n",
                         qmi_dms_firmware_image_type_get_string (image->type),
                         image_index);
